+++ noodles-bam/src/async/writer.rs
@@ -271,7 +271,7 @@ where
 
     writer.write_all(record.sequence()).await?;
     writer.write_all(record.quality_scores()).await?;
-    writer.write_all(&record.data).await?;
+    writer.write_all(record.data()).await?;
 
     Ok(())
 }