+++ noodles-bam/src/record/convert.rs
@@ -165,7 +165,7 @@ mod tests {
     fn build_record() -> Record {
         use sam::record::Flags;
 
-        use crate::record::{Cigar, QualityScores, Sequence};
+        use crate::record::{Cigar, Data, QualityScores, Sequence};
 
         Record {
             ref_id: 1,
@@ -180,10 +180,10 @@ mod tests {
             cigar: Cigar::from(vec![0x00000040]),    // 4M
             seq: Sequence::new(vec![0x18, 0x42], 4), // ATGC
             qual: QualityScores::from(vec![0x1f, 0x1d, 0x1e, 0x20]), // @>?A
-            data: vec![
+            data: Data::from(vec![
                 0x4e, 0x4d, 0x43, 0x00, // NM:i:0
                 0x50, 0x47, 0x5a, 0x53, 0x4e, 0x41, 0x50, 0x00, // PG:Z:SNAP
-            ],
+            ]),
         }
     }